pyproject.toml
[tool.poetry]
name = "my-elasticsearch"
version = "0.1.0"
description = ""
authors = ["Eduard Kaverinskyi <eduard.kaverinskyi@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "my_elasticsearch", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
pymongo = "^4.10.1"
pydantic = "^2.9.2"
python-dotenv = "^1.0.1"
pytest = "^8.3.3"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
